Remove unused rpath settings
Prior to this change if a module didn't link to anything under our root
we left the rpath from the build un-modified. This lead to rpaths
pointint to directories that may not exist.

This update will clear any un-used rpath settings, only modules that
require an rpath will have one set.